Moved classes responsible for sorting to one package
